build: rename patient-duplicates to find-patient-duplicates and move to ./admin folder
Ref. metriport/metriport-internal#704
